# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GNV OnBoard API",
    "version": "1.0.0",
    "description": "API REST versionnée : préfixe canonique `/api/v1`. Les mêmes routes sont montées aussi sous `/api/*` (alias rétrocompatibilité) — dans OpenAPI, seul le préfixe `/api/v1` est documenté.\n\n**Swagger UI** : `GET /api-docs` (dev par défaut ; prod si `SWAGGER_ENABLED=true`). **Spec JSON** : `GET /api-docs.json`.",
    "contact": {
      "name": "GNV Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    },
    {
      "url": "https://api.gnv-onboard.com",
      "description": "Serveur de production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authToken"
      }
    },
    "schemas": {
...
      "Restaurant": {
        "type": "object",
        "description": "Restaurant (champs localisés si `lang` est fourni)",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "menu": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "promotions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/restaurants": {
      "get": {
        "summary": "Liste des restaurants actifs",
        "tags": [
          "Restaurants"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "example": "fr"
            },
            "description": "Code langue pour traductions (ex. fr, en)"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK (MongoDB ou fallback fichier si DB hors ligne)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Restaurant"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/v1/restaurants/{id}": {
      "get": {
        "summary": "Détail d'un restaurant",
        "tags": [
          "Restaurants"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurant"
                }
              }
            }
          },
          "404": {
            "description": "Non trouvé"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    }
  }
}